Move controls under flowgraph folder
